Move logging.py into core
